Orginize code in app.
